Added newline at end of files and rename file
