--- src/components/Header/Header.css
@@ -1,4 +1,0 @@
-/* .navbar {
-  background-color: #5D5FEF;
-  color: #FFF;
-} */
