+++ src/components/Utils/PublicRoute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Navigate } from 'react-router-dom';
-import AuthService from '../../services/auth-service';
+import AuthService from '../../services/authService';
 
 export default function PublicRoute({ children }) {
   return AuthService.isLoggedIn()
