+++ src/routes/FavoritesPage/FavoritesPage.js
